Numerous Updates
Suppress unnecessary warnings
Update GEOROC list
Add stream_log for debugging
Update GEOROC download, import
Update pyrolite.util.time

+++ pyrolite/util/spatial.py
@@ -12,7 +12,7 @@ def piecewise(segment_ranges:list, segments=2, output_fmt=np.float):
     """
     outf = np.vectorize(output_fmt)
     if type(segments)==np.int:
-        segments = list(np.ones(len(segment_ranges)) * segments)
+        segments = list(np.ones(len(segment_ranges), dtype=np.int) * segments)
     else:
         pass
     seg_width = [(x2 - x1) / segments[ix]  # can have negative steps
